snippet infrastructure removed, no caching yet
