+++ capstone/config/celery.py
@@ -14,7 +14,6 @@ app = Celery('config', include=[
     'scripts.compress_volumes',
     'scripts.ice_volumes',
     'scripts.export',
-    'scripts.update_snippets',
     'scripts.refactor_xml',
     'scripts.make_pdf',
 ])
